Move AnnotatedText to its own module in lib
