+++ lib/src/AnnotatedText.ts
@@ -1,13 +1,32 @@
-import { AnnotatedText, ScopedText, TextAnnotation } from '@lib/protocol';
-export { AnnotatedText, ScopedText, TextAnnotation } from '@lib/protocol';
+export type TextDifference = "added"|"removed";
+
+export interface TextAnnotation {
+  /** the relationship this text has to the text of another state */
+  diff?: TextDifference,
+  /** what to display instead of this text */
+  substitution?: string,
+  /** the underlying text, possibly with more annotations */
+  text: string
+}
+
+export interface ScopedText {
+  /** A scope identifier */
+  scope: string,
+  /** Extra stuff */
+  attributes?: any,
+  /** the underlying text, possibly with more annotations */
+  text: AnnotatedText,
+}
+
+export type AnnotatedText = string | TextAnnotation | ScopedText | (string | TextAnnotation | ScopedText)[];
 
 
 export function isScopedText(text: AnnotatedText): text is ScopedText {
-  return text.hasOwnProperty('scope');
+  return text && text.hasOwnProperty('scope');
 }
 
 export function isTextAnnotation(text: AnnotatedText): text is TextAnnotation {
-  return typeof (text as any).text === 'string'
+  return text && typeof (text as any).text === 'string' && !text.hasOwnProperty('scope')
 }
 
 export function textToString(text: AnnotatedText) : string {
