+++ lib/src/index.ts
@@ -1,3 +1,4 @@
 export * from './protocol';
 export * from './settings';
 export * from './coqview';
+export * from './AnnotatedText';
