+++ server/src/util/AnnotatedText.ts
@@ -1,46 +1,12 @@
 import * as diff from 'diff';
 
-import { AnnotatedText, ScopedText, TextAnnotation } from '@lib/protocol';
-export { AnnotatedText, ScopedText, TextAnnotation } from '@lib/protocol';
+import { AnnotatedText, ScopedText, TextAnnotation, isScopedText, isTextAnnotation } from '@lib/AnnotatedText';
+export { AnnotatedText, ScopedText, TextAnnotation, isScopedText, isTextAnnotation, textToString, textToDisplayString } from '@lib/AnnotatedText';
 
 
 export interface Annotation {
   /** what to display instead of this text */
   substitution?: string,
-}
-
-export function isScopedText(text: AnnotatedText): text is ScopedText {
-  return text && text.hasOwnProperty('scope');
-}
-
-export function isTextAnnotation(text: AnnotatedText): text is TextAnnotation {
-  return text && typeof (text as any).text === 'string' && !text.hasOwnProperty('scope')
-}
-
-export function textToString(text: AnnotatedText) : string {
-  if(typeof text === 'string') {
-    return text;
-  } else if(text instanceof Array) {
-    return text.map(textToString).join('');
-  } else if(isScopedText(text)) {
-    return textToString(text.text);
-  } else {// TextAnnotation
-    return textToString(text.text);
-  }
-}
-
-export function textToDisplayString(text: AnnotatedText) : string {
-  if(typeof text === 'string') {
-    return text;
-  } else if(text instanceof Array) {
-    return text.map(textToDisplayString).join('');
-  } else if(isScopedText(text)) {
-    return textToDisplayString(text.text);
-  } else if(text.substitution) {// TextAnnotation
-    return textToDisplayString(text.substitution);
-  } else{// TextAnnotation
-    return text.substitution ? text.substitution : text.text;
-  }
 }
 
 export function textLength(text: AnnotatedText) : number {
